pyproject.toml
[tool.poetry]
name = "rag-batch-gemini-demo"
version = "0.1.0"
description = "Offline demo of batch Q&A with context caching"
authors = ["Shifat Bin Rahman <shifatrahman@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12" 
sentence-transformers = "^2.6.1"
faiss-cpu = "^1.7.4"
scipy = "1.10.1"
numpy = "1.24.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
